+++ tests/test_pipelines_training_dataset.py
@@ -2,56 +2,36 @@ import numpy as np
 import pytest
 import xarray as xr
 
-from fv3net.pipelines.create_training_data.pipeline import _create_train_cols
+from fv3net.pipelines.create_training_data.pipeline import _add_apparent_sources
 
 
 @pytest.fixture
 def test_training_raw_ds():
     centered_coords = {
-        "grid_yt": [1],
-        "grid_xt": [1],
-        "initialization_time": [
+        "initial_time": [
             np.datetime64("2020-01-01T00:00").astype("M8[ns]"),
             np.datetime64("2020-01-01T00:15").astype("M8[ns]"),
         ],
-        "forecast_time": [
-            np.datetime64("2020-01-01T00:00").astype("M8[ns]"),
-            np.datetime64("2020-01-01T00:15").astype("M8[ns]"),
-        ],
-    }
-    x_component_edge_coords = {
-        "grid_y": [0, 2],
-        "grid_xt": [1],
-        "initialization_time": [
-            np.datetime64("2020-01-01T00:00").astype("M8[ns]"),
-            np.datetime64("2020-01-01T00:15").astype("M8[ns]"),
-        ],
-        "forecast_time": [
-            np.datetime64("2020-01-01T00:00").astype("M8[ns]"),
-            np.datetime64("2020-01-01T00:15").astype("M8[ns]"),
-        ],
+        "forecast_time": np.array([0.0, 60.0, 120.0]).astype(np.dtype("<m8[s]")),
     }
 
-    # hi res changes by 0 K, C48 changes by 1 K
+    # hi res changes by 0 K, C48 changes by 1 K and 26 K
     T_da = xr.DataArray(
-        [[[[273.0, 274.0], [273.0, 274.0]]]],
-        dims=["grid_yt", "grid_xt", "initialization_time", "forecast_time"],
+        [[273.0, 274.0, 300.0], [273.0, 275.0, 310.0]],
+        dims=["initial_time", "forecast_time"],
         coords=centered_coords,
     )
 
-    u_da = xr.DataArray(
-        [[[[20, 30], [20.0, 30.0]]], [[[30, 40], [30.0, 40.0]]]],
-        dims=["grid_y", "grid_xt", "initialization_time", "forecast_time"],
-        coords=x_component_edge_coords,
+    return xr.Dataset({"air_temperature": T_da})
+
+
+def test__add_apparent_sources(test_training_raw_ds):
+    train_ds = _add_apparent_sources(
+        test_training_raw_ds,
+        tendency_tstep_onestep=1,
+        tendency_tstep_highres=1,
+        init_time_dim="initial_time",
+        forecast_time_dim="forecast_time",
+        var_source_name_map={"air_temperature": "dQ1"},
     )
-
-    ds = xr.Dataset({"T": T_da, "sphum": T_da, "u": u_da, "v": u_da})
-    return ds
-
-
-def test__create_train_cols(test_training_raw_ds):
-    train_ds = _create_train_cols(
-        test_training_raw_ds, cols_to_keep=["T", "dQ1", "dQU"]
-    )
-    assert pytest.approx(train_ds.dQ1.values) == -1.0 / (15.0 * 60)
-    assert pytest.approx(train_ds.dQU.values) == -10.0 / (15.0 * 60)
+    assert train_ds.dQ1.values == pytest.approx(1.0 / (15.0 * 60) - 26.0 / 60)