+++ workflows/one_step_jobs/_develop_logging_runfile.sh
@@ -1,0 +1,2 @@
+rm -rf test_outdir
+fv3run --runfile runfile_test_logging.py test_config.yml test_outdir